conform to bittensor api, update example
